Move teams, organizations and initiatives out of the apps directory.
