+++ mapstory/initiatives/templates/initiatives/card.html
