+++ mapstory/initiatives/templates/initiatives/detail.html
