+++ mapstory/initiatives/templates/initiatives/list.html
