+++ mapstory/initiatives/templates/initiatives/manager.html
