+++ mapstory/initiatives/templates/initiatives/mapstory_card.html
