+++ mapstory/initiatives/templates/initiatives/member_card.html
